Implement [skip ci] for Jenkins
GitHub Actions already respects this ([link](https://github.blog/changelog/2021-02-08-github-actions-skip-pull-request-and-push-workflows-with-skip-ci/)), this implements it for Jenkins as well. When `[skip ci]` is in the commit message of the head commit, the only things that will run are the sanity check + lint, so CI should be much quicker if someone manually sets this.

This is useful for PRs that need to get landed quickly and the outcome is pretty well known, such as reverts.